+++ extensions/iceworks-material-import/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 0.1.2
+
+- feat: quick activate iceworks-material-import extensions by editor context
+
 ## 0.1.1
 
 - style: better style of the webview