+++ extensions/iceworks-style-helper/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 0.3.1
+
+- Support ICE css module style preview and completion.
+- Support className preview, definition and completion.
+
 ## 0.3.0
 
 Add Sass Variables Helper: